Replace grunt with Makefile & upgrade manifest version
Has been a while since I ran this, and a lot of the node plugins used
were outdated, and many of them had security issues. Instead of fixing
all that dependency tree, opted out for the much simpler approach of
having a Makefile. Grut was an overkill anyway, and most of the stuff
weren't really necessary (e.g. ugligying and minifying js, since they
are only a few lines anyway).

Upgraded `manifest_version` to 3.